// frontend/ticket-booking-web/src/app/api/model/create-event-request.ts
/**
 * Ticket Booking Platform API - v1
 *
 * Contact: m1musbal@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EventCategory } from './event-category';


export interface CreateEventRequest { 
    title: string;
    description: string;
    category: EventCategory;
    startsAt: string;
    endsAt: string;
    locationId: string;
}
export namespace CreateEventRequest {
}